nextbike-scraper/requirements.txt
requests==2.32.4
python_dotenv==1.1.1
pymongo==4.13.2
